[ UPDATE-postpage ] postpage에서 각 부분별로 컴포넌트를 분리하라 - input / image / button 컴포넌트로 분리하여 각각 컴포넌트에서 할 일을 나눠 하였다 - input컴포넌트에서는 사용자 데이터를 받아 contextAPI로 데이터를 업데이트하였고 - image컴포넌트에서는 input file을 URL로 받아 사용자에게 미리 볼 수 있게 하고, contextAPI로 file데이터를 저장하였다 - button컴포넌트에서는 위 컴포넌트에서 contextAPI에 업데이트 한 데이터를 가지고와 서버로 요청하였다. - 이때 S3로 업로드하는 부분과 APi요청하는 부분은 util과 hooks으로 생성하여 함수를 호출하여 처리하였다
