+++ app/postPage/components/ImageComponent.jsx
@@ -1,28 +1,48 @@
 "use client";
 import React, { useContext, useState } from "react";
-import { ImageDataContext } from "@/app/context/ImageContext";
-import styles from "../postPage.module.css";
 import Image from "next/image";
+import { ImageDataContext } from "@/app/context/ImageContext";
 import { imagePreviewUtil } from "@/utils/imageUpload";
+import styles from "../postPage.module.css";
 
-export default function ImageComponent() {
+const EditDefaulImage = (imageDefault) => {
+  return (
+    <label htmlFor="file" className={styles.imageBasicLabel}>
+      <Image src={imageDefault} alt="default-image" width={490} height={400} />
+    </label>
+  );
+};
+
+const PostIconImage = () => {
+  return (
+    <label htmlFor="file" className={styles.imageBasicLabel}>
+      <Image
+        src="/icons/gray_image_add.svg"
+        alt="image-add-icon"
+        width={100}
+        height={100}
+      />
+      <p>오늘의 툰을 올려주세요</p>
+    </label>
+  );
+};
+
+export default function ImageComponent({ type, imageDefault}) {
   const { setImageState } = useContext(ImageDataContext);
   const [previewImage, setPreviewImage] = useState("");
 
+  // image 미리보기
   const handleImagePreview = async (e) => {
     const file = e.target.files?.[0];
     setImageState({ imageFile: file });
-
-    // image 미리보기
     if (file) {
       const preview = await imagePreviewUtil(file);
       setPreviewImage(preview);
     }
   };
-  
+
   return (
     <div className={styles.postImageBox}>
-      {/* input=file style변경 */}
       <div className={styles.labelBox}>
         {/* 실제 image업로드 input */}
         <input
@@ -32,22 +52,11 @@ export default function ImageComponent() {
           accept="image/*"
           onChange={handleImagePreview}
         />
+        {/* 기본input 대신 보여지는 Image / ICON */}
         {!previewImage ? (
-          <>
-            {/* input=file 기본style 대신 */}
-            <label htmlFor="file" className={styles.imageBasicLabel}>
-              <Image
-                src="/icons/gray_image_add.svg"
-                alt="image-add-icon"
-                width={100}
-                height={100}
-              />
-              <p>오늘의 툰을 올려주세요</p>
-            </label>
-          </>
+          type !== 'edit' ? PostIconImage() : EditDefaulImage(imageDefault)
         ) : (
           <div className={styles.imageEditLabelBox}>
-            {/* preview이미지 있을 때 img수정 input */}
             <label htmlFor="file" className={styles.imageEditLabel}>
               <p>이미지 변경</p>
             </label>
@@ -60,6 +69,7 @@ export default function ImageComponent() {
           </div>
         )}
       </div>
+      
       {/* Preview이미지 보여지는 부분 */}
       {previewImage && (
         <Image
